refactor: раализовать отдельные классы-конвертеры для каждой сущности, в которых будут методы для конвертации в различные дто и обратно

+++ src/main/java/ru/yandex/practicum/filmorate/dto/FilmDto.java
@@ -19,5 +19,5 @@ public final class FilmDto {
     private final String releaseDate;
 
     @Positive(message = "Продолжительность фильма должна быть положительной")
-    private final int duration;
+    private final long duration;
 }
